fl-adni-classification/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adni1"
description = "Centralized 3-way classification (CN vs MCI vs AD) of ADNI 3D MRI data"
readme = "README.md"
version = "1.0.1"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Tin Hoang" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "torch>=2.0.0",
    "monai>=1.2.0",
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.12.0",
    "tqdm>=4.62.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "nibabel>=5.2.1",
    "wandb>=0.15.0",
    "filelock>=3.15.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "mypy>=1.15.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
]

# Tools settings: pre-commit ----------------------------------------------------------------------
[tool.ruff]
# https://docs.astral.sh/ruff/configuration/
line-length = 120
indent-width = 4

# Exclude folders/files from linting
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "*.egg-info",
    "build",
    "dist",
    "docs",
    "configs",
    "data",
    "scripts",
    "models",
    "notebooks",
    "scripts",
    "outputs",
    "wandb",
    # Add your custom folders here
    # "docs",
    # "scripts",
    # "tests",
]

[tool.ruff.lint]
# see: https://docs.astral.sh/ruff/rules
select = [
     "F", # Pyflakes
     "E", # Pycodestyle
     "W", # Pycodestyle
     "I", # Isort
     "B", # Flake8-bugbear
     "C4", # Flake8-comprehensions
]
ignore = [
    "E731", # Do not assign a lambda expression, use a def
    "E741", # Do not use variables named 'I', 'O', or 'l'
    "C408", # Unnecessary <dict/list/tuple> call - rewrite as a literal
    "W191", # Indentation contains tabs (incompatible with `format.indent-style="tab"`)
    "C401", # Unnecessary generator (rewrite as a `set` comprehension)
    "A005", # Unnecessary `any` type
    "ANN401",  # any-type
    "B008", # Unnecessary `self` parameter
    "C901",  # complex-structure
    "EM101",  # Raw string in exception
    "EM102",  # F-string in exception
    "FA102",  # Future required type annotation
    "FIX002",  # Line contains todo
    "N815",  # Mixed-case variable in class scope
    "PLR0911",  # Too many return statements
    "PLR0913",  # Too many arguments
    "RUF001",  # Ambiguous unicode character string
    "S105", # Missing docstring in magic method
]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
explicit_package_bases = true
namespace_packages = true
exclude = [
    "scripts/.*",
    "3D_CNN_pretrained_model/.*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=adni_classification --cov-report=term-missing"

[tool.bandit]
exclude_dirs = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "docs",
    "configs",
    "data",
    "scripts",
    "models",
    "notebooks",
    "tests",
    "outputs",
    "wandb"
]
# Skip common ML/research patterns that are generally acceptable
skips = [
    "B101",  # assert_used - common in testing/research
    "B110",  # try_except_pass - common in cleanup code
    "B112",  # try_except_continue - common in ML loops
    "B404",  # subprocess_import - needed for system interaction
    "B507",  # ssh_no_host_key_verification - acceptable in research clusters
    "B601",  # shell_injection_process_args
    "B602",  # subprocess_popen_with_shell_equals_true - needed for some operations
    "B614",  # pytorch_load_save - acceptable with trusted checkpoints
]

[tool.hatch.build.targets.wheel]
packages = ["adni_classification"]

[project.urls]
"Homepage" = "https://github.com/tin-hoang/fl-adni-classification"
"Bug Tracker" = "https://github.com/tin-hoang/fl-adni-classification/issues"

